+++ homeassistant/components/roku/helpers.py
@@ -1,6 +1,21 @@
 """Helpers for Roku."""
 from __future__ import annotations
 
+from collections.abc import Awaitable, Callable, Coroutine
+from functools import wraps
+import logging
+from typing import Any, TypeVar
+
+from rokuecp import RokuConnectionError, RokuConnectionTimeoutError, RokuError
+from typing_extensions import Concatenate, ParamSpec
+
+from .entity import RokuEntity
+
+_LOGGER = logging.getLogger(__name__)
+
+_T = TypeVar("_T", bound=RokuEntity)
+_P = ParamSpec("_P")
+
 
 def format_channel_name(channel_number: str, channel_name: str | None = None) -> str:
     """Format a Roku Channel name."""
@@ -8,3 +23,43 @@ def format_channel_name(channel_number: str, channel_name: str | None = None) ->
         return f"{channel_name} ({channel_number})"
 
     return channel_number
+
+
+def roku_exception_handler(
+    func: Callable[Concatenate[_T, _P], Awaitable[None]],  # type: ignore[misc]
+) -> Callable[Concatenate[_T, _P], Coroutine[Any, Any, None]]:  # type: ignore[misc]
+    """Decorate Roku calls to handle Roku exceptions."""
+
+    @wraps(func)
+    async def wrapper(self: _T, *args: _P.args, **kwargs: _P.kwargs) -> None:
+        try:
+            await func(self, *args, **kwargs)
+        except RokuConnectionError as error:
+            if self.available:
+                _LOGGER.error("Error communicating with API: %s", error)
+        except RokuError as error:
+            if self.available:
+                _LOGGER.error("Invalid response from API: %s", error)
+
+    return wrapper
+
+
+def roku_exception_handler_no_timeout(
+    func: Callable[Concatenate[_T, _P], Awaitable[None]],  # type: ignore[misc]
+) -> Callable[Concatenate[_T, _P], Coroutine[Any, Any, None]]:  # type: ignore[misc]
+    """Decorate Roku calls to handle Roku exceptions."""
+
+    @wraps(func)
+    async def wrapper(self: _T, *args: _P.args, **kwargs: _P.kwargs) -> None:
+        try:
+            await func(self, *args, **kwargs)
+        except RokuConnectionTimeoutError:
+            return
+        except RokuConnectionError as error:
+            if self.available:
+                _LOGGER.error("Error communicating with API: %s", error)
+        except RokuError as error:
+            if self.available:
+                _LOGGER.error("Invalid response from API: %s", error)
+
+    return wrapper
